Export CMake targets from top-level directory
At some point I think I have seen that the location in which that install command is triggered affects the location of the installed file. For us, doing it from the top-level doesn't change anything, so why not play it safe? Also, this is the same that Gloo does (and Gloo also tries to keep most of its CMake logic inside its gloo/ subdir).

ghstack-source-id: 2e418e19ce23198fb85d725ffa9a73d46108f227
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/194