Export submodule libuv inside TensorPipe's targets
That target is a hack we need in order to be able to export the libuv target. I think it would be weird for users to install TensorPipe and then find a weird thing called libuv-unofficial installed on their machine. Why not keep it under our umbrella to make attribution easier? A later PR in this stack will also change the name of the target itself (currently `_uv_a`) as well.

ghstack-source-id: a6f866213d482e37482bf7c9a161cacb5cc3b13c
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/195