Allow to build TensorPipe as a static lib
Sometimes (I'm looking at you, PyTorch) users include TensorPipe as a submodule in a project, where they want to build the top-level library as shared, and thus BUILD_SHARED_LIBS is enabled, but they want to build and link TensorPipe *statically* into that shared library. So we should offer a way for users to override the default behavior. (However, even when TensorPipe is built statically, if the final result is dynamic we must build position-independent code).

ghstack-source-id: 08c37335107b10d7a1fd618e0a62d23b6679f8f4
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/198