Allow downstream projects to customize install location
It makes perfect sense for us to install in the standard location for libraries in the system. However, this sometimes turns out to be `lib64`, while PyTorch hardcodes such a location as `lib`. Not sure it matters, and if anything it should be PyTorch to fix this, but let's be nice and allow users to specify a different location if they want.

ghstack-source-id: cf53cb0983f847f9ced34cda4134c7bcb86e17d4
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/197